Refactor CompilationExtensions and MemoryExtensionsBuilder; improve type handling and readability
Enhanced type handling for unsigned types in CompilationExtensions and improved readability in MemoryExtensionsBuilder.
This change simplifies the logic for creating vectors and ensures proper casting for various types, making the code more maintainable and easier to understand. 🔧✨

+++ Vectorize/ConstExpr.Sample/Test.cs
@@ -59,17 +59,17 @@ public static class Test
 		return nameof(Test);
 	}
 
-	public static ICustomCollection<int> Range(int count)
+	public static ICustomCollection<byte> Range(int count)
 	{
 		var random = new Random();
-		var result = new List<int>(count);
+		var result = new List<byte>(count);
 
 		for (var i = 0; i < count; i++)
 		{
-			result.Add(random.Next(5));
+			result.Add((byte)random.Next(5));
 		}
 
-		return result.OrderBy(o => o) as ICustomCollection<int>;
+		return result.OrderBy(o => o) as ICustomCollection<byte>;
 	}
 
 	public static ICustomCollection<string> Split(string value, char separator)
